% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efectos.R
\name{alleffect}
\alias{alleffect}
\title{Función que hace la grafica de efectos de todas las variables de un modelo.
Recide como parametros:}
\usage{
alleffect(model, alphas)
}
\arguments{
\item{model}{- objeto del tipo Dvine}

\item{alphas}{- vector de double. Vector que contiene los niveles con los que
se va a hacer la regresión cuantil.}
}
\value{
None
}
\description{
Función que hace la grafica de efectos de todas las variables de un modelo.
Recide como parametros:
}
