% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion.R
\name{constCop}
\alias{constCop}
\title{Funcion que contruye el ajuste de copulas. Recibe como parametros:}
\usage{
constCop(formula, data, estima = FALSE, criterio = "BIC")
}
\arguments{
\item{formula}{- la formula a ajustar}

\item{data}{- dataframe con los datos a ajustes (no tienen que estar en
escala u)}

\item{estima}{- boolean que indica, si se ajusta el modelo con la formula (FALSE)
o se contruye el primer árbol maximizando la correlacion de wolff (TRUE)}

\item{criterio}{- El criterio con el que se ajustan las copulas 'BIC', 'AIC' o 'loglikehood'}
}
\value{
Retorna un objeto del tipo model Dvine
}
\description{
Funcion que contruye el ajuste de copulas. Recibe como parametros:
}
