% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion.R
\name{corSpear}
\alias{corSpear}
\title{Función que saca la correlacion de Spearman.
Recibe como argumento:}
\usage{
corSpear(datos)
}
\arguments{
\item{datos}{- un data frame con todas las variables que se
les quiere sacar la correlacion de Spearman.}
}
\value{
Retorna la matriz de correlacion.
}
\description{
Función que saca la correlacion de Spearman.
Recibe como argumento:
}
