% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion.R
\name{corSWEC}
\alias{corSWEC}
\title{Función que saca la correlacion wolf con la funcion que R tiene
implementada en el paquete copBasic. Para hacer el arbol o el camino
base de la D-vine.Recibe como argumento:}
\usage{
corSWEC(datos)
}
\arguments{
\item{datos}{un data frame con todas las variables que se
les quiere sacar la correlacion de igma de Wolf and Schewiezer.}
}
\value{
Retorna la matriz de correlacion.
}
\description{
Función que saca la correlacion wolf con la funcion que R tiene
implementada en el paquete copBasic. Para hacer el arbol o el camino
base de la D-vine.Recibe como argumento:
}
