% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraphs.R
\name{diagonalGraf}
\alias{diagonalGraf}
\title{Funcion que hace graficos de dispersion
Hacemos las graficas de las diagonales en la
Con las tablas de dispersion en la triangular superior.
Recibe como parametro:}
\usage{
diagonalGraf(model)
}
\arguments{
\item{model}{- un objeto del tipo D-vine}
}
\value{
None
}
\description{
Funcion que hace graficos de dispersion
Hacemos las graficas de las diagonales en la
Con las tablas de dispersion en la triangular superior.
Recibe como parametro:
}
