% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion.R
\name{escalaU}
\alias{escalaU}
\title{Funcion que convierte variables a su escala U. Recibe como argumentos:}
\usage{
escalaU(datos)
}
\arguments{
\item{datos}{- dataframe donde cada columna corresponde a una variable}
}
\value{
Retorna el dataframe con los datos en escala u
}
\description{
Funcion que convierte variables a su escala U. Recibe como argumentos:
}
