% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatsCopula}
\alias{heatsCopula}
\title{Funcion que hace la grafica de cualquier copula. La
diferencia con la funcion anterior la selección del data frame
de entrada. Pues esta depende del nivel de la copula.
Recibe como argumento:}
\usage{
heatsCopula(model, copName)
}
\arguments{
\item{model}{- un objetivo del tipo D-vine}

\item{copName}{- string que tiene el nombre de la copula, ejemplo 'C12'}
}
\description{
Funcion que hace la grafica de cualquier copula. La
diferencia con la funcion anterior la selección del data frame
de entrada. Pues esta depende del nivel de la copula.
Recibe como argumento:
}
