% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatVine}
\alias{heatVine}
\title{Funcion que hace el graficos de todo el modelo de la D-vine.
Recibe como parametros:}
\usage{
heatVine(model)
}
\arguments{
\item{model}{- un objeto del tipo del D-vine}
}
\value{
none
}
\description{
Funcion que hace el graficos de todo el modelo de la D-vine.
Recibe como parametros:
}
