% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efectos.R
\name{plot_effects}
\alias{plot_effects}
\title{Funcion que hace la grafica de effectos de la variable (solo una)
deseadas. Recibe como argumentos:}
\usage{
plot_effects(model, variable, alpha)
}
\arguments{
\item{model}{- un objeto del tipo Dvine}

\item{variable}{- string con la variable de cual se quiere ver los efectos}

\item{alpha}{- vector de double, el nivel alpha con el cual se va a hacer
la regresión cuantil}
}
\value{
Retorna un 1, si la función tiene éxito.
}
\description{
Funcion que hace la grafica de effectos de la variable (solo una)
deseadas. Recibe como argumentos:
}
