% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{regresionPred}
\alias{regresionPred}
\title{Funcion que hace la regresion de la variable obser, usando
las copulas generadas. Recibe como argumento:}
\usage{
regresionPred(model, dataPred, alpha)
}
\arguments{
\item{model}{- un objeto del tipo de Dvine}

\item{dataPred}{- data frame con las observaciones que se les quiere
hacer prediccion, no tiene que estar en escala U.}

\item{alpha}{- nivel al cual se va a hacer la regresión
NOTA: Obser deben tener los mismos nombres que en las variables en escala U}
}
\description{
Funcion que hace la regresion de la variable obser, usando
las copulas generadas. Recibe como argumento:
}
