% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efectos.R
\name{showCopula}
\alias{showCopula}
\title{Funcion que grafica las copulas. La primera grafica
las curvas de nivel, la segunda es la grafica de la diagonal.
Recibe como parametros:}
\usage{
showCopula(model, copName, col = viridis::inferno(20))
}
\arguments{
\item{model}{- un objetivo del D-vine}

\item{copName}{- string, con el nombre de la copula, ejemplo 'C12'}

\item{col}{- vector de colores para la curva de nivel}
}
\description{
Funcion que grafica las copulas. La primera grafica
las curvas de nivel, la segunda es la grafica de la diagonal.
Recibe como parametros:
}
