% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efectos.R
\name{ShowVine}
\alias{ShowVine}
\title{Función que hace las grafias de showCopula para todo
el modelo Dvine. Recibe como parametros:}
\usage{
ShowVine(model, col = inferno(20))
}
\arguments{
\item{model}{- un objeto de tipo D-vine}

\item{col}{- gamma de colores para las curvas de nivel}
}
\value{
None
}
\description{
Función que hace las grafias de showCopula para todo
el modelo Dvine. Recibe como parametros:
}
