% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresion.R
\name{testCop}
\alias{testCop}
\title{Funcion que imprime los test de independencia de cada copula.
Recibe como parametro:}
\usage{
testCop(model)
}
\arguments{
\item{model}{- Un objetivo del tipo model Dvine}
}
\value{
None
}
\description{
Funcion que imprime los test de independencia de cada copula.
Recibe como parametro:
}
