% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{W}
\alias{W}
\title{Funcion que calcula la copula inferior (Bidimensional)
Recibe como parametro:}
\usage{
W(U)
}
\arguments{
\item{U}{- data.frame dimensional}
}
\value{
Retorma un dataframe unidimensional
}
\description{
Funcion que calcula la copula inferior (Bidimensional)
Recibe como parametro:
}
